authenticity-validator/ai-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
